init python3 conversion

+++ obfsproxy/common/aes.py
@@ -19,7 +19,7 @@ class AES_CTR_128(object):
         assert(len(key) == 16)
         assert(len(iv) == 16)
 
-        self.ctr = Counter.new(128, initial_value=long(iv.encode('hex'), 16),
+        self.ctr = Counter.new(128, initial_value=int(iv.encode('hex'), 16),
                                allow_wraparound=counter_wraparound)
         self.cipher = AES.new(key, AES.MODE_CTR, counter=self.ctr)
